Wrote input/output front-end
go to /admin in browser (be sure cgi is enabled for execution in /cgi-bin, and that all files in /cgi-bin are +x, and output.txt is +w), the page should load test content in the editor, then type what you want, submit, smile! reload page and what you submitted should show up again.